Added CLI specs for reading and opening gem.
These are basic tests to ensure the open and read commands can be
messaged properly. We might be able to improve upon behavior in the
future.

A fake gem of `gemsmith_test` is used for these specs in order to not
cause confusion with local development. This avoids opening your
default editor and browser.