Fixed CLI spec template to account for special characters in gem labels.
Necessary to prevent specs failures when generating a new gem where the
gem label has special characters in it. Example:

    module ExamplePlus
      module Identity
        def self.name
          "gemsmith"
        end

        def self.label
          "Example+"
        end

        def self.version
          "1.0.0"
        end
      end
    end

The use of `ExamplePlus::Identity.label` in a regular expresssion would
have failed the spec because the plus (i.e. `+`) has special meaning.
This change ensures that doesn't happen in order to not confuse gem
authors.