Updated readme builder to handle gem setup and security
Necessary since this used to be handled by the Rubysmith project and is now accounted for in this project.

Updated the RuboCop configuration to ignore long methods due to some of the longer ASCII Doc and Markdown content. Splitting this up further or adding to a separate file wouldn't be of much benefit in terms of maintenance.

Milestone: minor