Fixed CLI builder with Refinements disabled
Ensures both of these files are built properly when Refinements are disabled. Granted, anyone _not_ using the Refinements gem has more work ahead of them but the specs pass upon initial generation now.

This required updating the RSpec fixtures to account for this change as well as inform RuboCop that we want to ignore layout issues with the proof.

Milestone: patch