Updated gem specification template to allow current version or higher
In the past, this was always locked to the current minor version and higher with the intent that you put in the effort to support a major release since a major Ruby release could introduce breaking changes. The other reason is that Ruby doesn't follow link:https://semver.org[Semantic Versioning] which is another source of issues. This practice, unfortunately, causes end-of-year issues when a new Ruby version drops because you have to:

. Release a version of all gems with a lower and upper version range in order to support the end of year version transition. Example `spec.required_ruby_version = ">= 3.4", "<= 3.5"`.
. Then, in the New Year, release another version of all gems using the new minor version lock. Example: `spec.required_ruby_version = "~> 3.5"`.
. Then repeat this process during the last couple of months before end of the next year.

The intent was to always support the latest version of Ruby, switch to the newer syntax, and push people to keep moving Ruby forward. Unfortunately, if you don't keep up with the new Ruby version changes, you prevent people from upgrading or playing with release candidates. So this change removes the upper bound in order to make this process easier and encourage folks to upgrade freely.

Milestone: minor